restructured project file structure
