+++ src/App.css
@@ -1,9 +1,5 @@
 .container {
     margin: 0;
     text-align: center;
-<<<<<<< HEAD
-}
-=======
     width: 100%;
 }
->>>>>>> Testing
