--- src/Pages/Home.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import Menu from '../Components/Menu';
-
-function Home() {
-  return (
-    <>
-        <Menu />
-    </>
-  )
-}
-
-export default Home;
