+++ src/components/Menu.jsx
@@ -1,18 +1,18 @@
-import React from 'react';
-import './Menu.css';
-import menuBanner from '../Images/menuBanner.png'
+import React from "react";
+import "../styles/Menu.css";
+import menuBanner from "../images/menuBanner.png";
 
 function Menu() {
   return (
     <>
-        <div className='menu-container'>
-            <div className='menu-banner'></div>
-            <div className='menu-list'>
-                <h1>Hewwo</h1>
-            </div>
+      <div className="menu-container">
+        <div className="menu-banner"></div>
+        <div className="menu-list">
+          <h1>Hewwo</h1>
         </div>
+      </div>
     </>
-  )
+  );
 }
 
-export default Menu
+export default Menu;
