+++ src/pages/Home.jsx
@@ -2,6 +2,8 @@ import React, { useEffect, useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import { UserContext } from "../App";
 
+import Menu from "../components/Menu";
+
 function Home() {
   const navigate = useNavigate();
   const [userId, setUserId] = useContext(UserContext);
@@ -14,7 +16,7 @@ function Home() {
 
   return (
     <div>
-      <h1>This is the Home Page</h1>
+      <Menu />
     </div>
   );
 }
