+++ src/styles/Navbar.css
@@ -1,4 +1,4 @@
-*{
+* {
     box-sizing: border-box;
 }
 
@@ -32,7 +32,7 @@ body {
     background-color: #555;
 }
 
-.navbar a{
+.navbar a {
     color: inherit;
     text-decoration: none;
     height: 100%;
@@ -41,6 +41,6 @@ body {
     padding: .25rem;
 }
 
-.navbar a:hover{
+.navbar a:hover {
     background-color: #777;
 }
